# swagger/swagger.yaml
swagger: '2.0'
info:
  version: 1.0.0
  title: Smart Serve
  description: Smart Serve API Documents
  contact:
    name: Smart Serve
    url: 'https://www.mobileprogramming.com/'
host: '104.211.74.123:3090'
basePath: /
schemes:
  - http
  - https
securityDefinitions:
  JWT:
    type: apiKey
    name: x-access-token
    in: header
    description: Enter valid API authorization token
...
paths:
...
  '/restaurant/{restaurantId}':
    delete:
      tags:
        - Restaurant Controller
      summary: This API delete particular restaurant
      operationId: deleteRestaurant
      produces:
        - application/json
      parameters:
        - name: restaurantId
          in: path
          type: string
          required: true
      responses:
        '200':
          description: Success
          schema:
            $ref: '#/definitions/SuccessResponse'
        default:
          description: Error
          schema:
            $ref: '#/definitions/ErrorResponse'
  '/restaurant/{restaurantId}/products':
    get:
      tags:
        - Restaurant Controller
      summary: This API return products list of particlar restaurant
      operationId: getProducts
      produces:
        - application/json
      parameters:
        - name: restaurantId
          in: path
          type: string
          required: true
        - name: veg
          in: query
          type: boolean
        - name: name
          in: query
          type: string
      responses:
        '200':
          description: Success
          schema:
            $ref: '#/definitions/SuccessResponse'
        default:
          description: Error
          schema:
            $ref: '#/definitions/ErrorResponse'
...
definitions:
  SuccessResponse:
    type: object
    properties:
      status:
        type: string
        example: success
      message:
        type: string
        example: Successfully job done.
      data:
        type: object
        description: Required result object or array of data.
  ErrorResponse:
    type: object
    properties:
      status:
        type: string
        example: error
      message:
        type: string
        description: Error message.
        example: Network error found. Please try after some time.
...